% Generated by roxygen2 (4.0.2): do not edit by hand
\name{unbind}
\alias{unbind}
\title{Save each column of a data frame to a separate file}
\usage{
unbind(x, .destdir, .compress = TRUE, .parallel = FALSE)
}
\arguments{
\item{x}{A data frame}

\item{.destdir}{Where to save the files.  For proper operation, it
cannot contain other files with \code{.rds} extension.}

\item{.compress}{Passed as \code{compress} to \code{\link[base]{save}}}

\item{.parallel}{If true, parallelize with \code{\link[parallel]{mclapply}}}
}
\description{
This function save the data for each column of a data frame to a separate
RDS file.  If the target directory does not exist, it is created, including
intermediate directories.  Compression is optional.
}
\details{
Use this function in a helper script for your package to add the files needed
for column-wise loading this data set, e.g., in the \code{data-raw} directory.
According to the "Writing R Extensions" manual,
"the convention has grown up to use directory \code{inst/extdata} for [other
data files needed by the package]"; use a subdirectory for each data frame.
}
\references{
\href{http://r-pkgs.had.co.nz/data.html}{R packages: External data}

\href{http://cran.r-project.org/doc/manuals/R-exts.html}{Writing R extensions}
}

