% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{iris_unbound}
\alias{iris_unbound}
\title{The Iris data, saved as one file per column}
\format{\preformatted{Class 'try-error'  atomic [1:1] Error in splice("inst/extdata/iris") : 
  Source directory inst/extdata/iris does not appear to contain .rds files.

  ..- attr(*, "condition")=List of 2
  .. ..$ message: chr "Source directory inst/extdata/iris does not appear to contain .rds files."
  .. ..$ call   : language splice("inst/extdata/iris")
  .. ..- attr(*, "class")= chr [1:3] "simpleError" "error" "condition"
}}
\usage{
iris_unbound
}
\description{
This is the same data as \code{\link[datasets]{iris}}, saved in
\code{inst/extdata} using one file per column.
}
\examples{
head(iris_unbound[1:3])
}
\keyword{datasets}

