% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{iris_unbound}
\alias{iris_unbound}
\title{The Iris data, saved as one file per column}
\format{\preformatted{List of 1
 $ envir:<environment: 0x39fe198> 
 - attr(*, ".ColNames")= chr(0) 
 - attr(*, ".Path")= chr "inst/extdata/iris"
 - attr(*, ".Files")= chr(0) 
 - attr(*, "class")= chr "dfsplice"
}}
\usage{
iris_unbound
}
\description{
This is the same data as \code{\link[datasets]{iris}}, saved in
\code{inst/extdata} using one file per column.
}
\examples{
head(iris_unbound[1:3])
}
\keyword{datasets}

