% Generated by roxygen2 (4.0.2): do not edit by hand
\name{splice}
\alias{splice}
\title{Loads column-wise data}
\usage{
splice(path)
}
\arguments{
\item{path}{Source directory as passed to \code{\link{unbind}}}
}
\value{
An object of class \code{dfunbind}; extract columns by using \code{[},
  \code{[[} or \code{$}
}
\description{
This function loads the data for each column of a data frame from a separate
RDS file, as created by \code{\link{unbind}}.
}
\details{
If this function is called from a namespace (e.g., when loading a package)
and \code{path} is a relative directory, it is assumed to be relative to the
installation path of the package (via \code{\link[base]{system.file}}).
}

