% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/splice.R
\name{splice}
\alias{isplice}
\alias{splice}
\title{Loads column-wise data}
\usage{
splice(path)

isplice(dataset, package = packageName())
}
\arguments{
\item{path}{Source directory as passed to \code{\link{unbind}}}

\item{dataset}{The name of the directory under \code{inst/extdata} where the
data is stored.}

\item{package}{The name of the package where the data is stored.  The default
loads data from the current package.}
}
\value{
An object of class \code{dfsplice}; extract columns by using \code{[},
  \code{[[} or \code{$}
}
\description{
These functions load the data for each column of a data frame from a separate
RDS file, as created by \code{\link{unbind}}.
}
\details{
Normally, you want to use \code{isplice} with the name of the directory
(under \code{extdata}) where the data is located. The \code{splice} function
allows specifying a path in the file system; it is called internally from
\code{isplice}.
}

