% Generated by roxygen2 (4.0.2): do not edit by hand
\name{unbind_one}
\alias{unbind_one}
\title{Save each column of a data frame to a separate file}
\usage{
unbind_one(x, name, .destdir_pattern = "inst/extdata/\%s", .compress = TRUE)
}
\arguments{
\item{x}{A data frame}

\item{name}{A variable name under which this data frame should be accessible
later}

\item{.destdir_pattern}{Where to save the files.  The default allows
including the files in a package.}

\item{.compress}{Passed as \code{compress} to \code{\link[base]{save}}}
}
\description{
This function save the data for each column of a data frame to a separate
RData file.  If the target directory does not exist, it is created, including
intermediate directories.  Compression is optional.
}

